:white_check_mark: Now detect Vite or Nuxt

+++ cli/cli-vue/deleteVueComponent.js
@@ -1,6 +1,9 @@
 const fs = require("fs");
 const inquirer = require("../utils/inquirer");
 
+let componentFolderVite = fs.existsSync("./src/components");
+let componentFolderNuxt = fs.existsSync("./components");
+
 async function deleteVueComponent() {
   try {
     await inquirer
@@ -12,11 +15,27 @@ async function deleteVueComponent() {
         },
       ])
       .then((answers) => {
-        let componentFile = `./components/${answers.componentName}.vue`;
+        if (componentFolderVite) {
+          console.log("Removing component");
+          setTimeout(() => {
+            let componentFile = `./src/components/${answers.componentName}.vue`;
 
-        fs.rm(componentFile, { recursive: true }, () =>
-          console.log(`Component "${answers.componentName}" deleted! 🗑️`)
-        );
+            fs.rm(componentFile, { recursive: true }, () =>
+              console.log(`Component "${answers.componentName}" deleted! 🗑️`)
+            );
+          }, 2000);
+        }
+
+        if (componentFolderNuxt) {
+          console.log("Removing component");
+          setTimeout(() => {
+            let componentFile = `./components/${answers.componentName}.vue`;
+
+            fs.rm(componentFile, { recursive: true }, () =>
+              console.log(`Component "${answers.componentName}" deleted! 🗑️`)
+            );
+          }, 2000);
+        }
       });
   } catch (error) {
     console.error(error.message);
